requirements.txt
streamlit
pandas
pdfplumber
python-docx
reportlab
openpyxl
